glib: Auto generate various win32 functions
