+++ glib/Gir.toml
@@ -56,6 +56,41 @@ manual = [
 name = "GLib.*"
 status = "generate"
     [[object.function]]
+    name = "win32_locale_filename_from_utf8"
+    cfg_condition = "windows"
+    ignore = true # TODO: implement me
+    [[object.function]]
+    name = "win32_get_windows_version"
+    cfg_condition = "windows"
+    [[object.function]]
+    name = "win32_getlocale"
+    cfg_condition = "windows"
+    [[object.function]]
+    name = "win32_check_windows_version"
+    cfg_condition = "windows"
+    [[object.function]]
+    name = "win32_error_message"
+    cfg_condition = "windows"
+    [[object.function]]
+    name = "win32_ftruncate"
+    cfg_condition = "windows"
+    ignore = true # TODO: implement me
+    [[object.function]]
+    name = "win32_get_command_line"
+    cfg_condition = "windows"
+    [[object.function]]
+    name = "win32_get_package_installation_directory"
+    cfg_condition = "windows"
+    ignore = true # TODO: implement me
+    [[object.function]]
+    name = "win32_get_package_installation_directory_of_module"
+    cfg_condition = "windows"
+    manual = true
+    [[object.function]]
+    name = "win32_get_package_installation_subdirectory"
+    cfg_condition = "windows"
+    ignore = true # TODO: implement me
+    [[object.function]]
     pattern = "(assertion_message_cmpint|assert_warning|assertion_message|assertion_message_cmpnum|assertion_message_cmpstr|warn_message|return_if_fail_warning)"
     ignore = true # Not useful assertions functions
     [[object.function]]
@@ -886,3 +921,8 @@ status = "generate"
 name = "GLib.Uri"
 status = "generate"
 concurrency = "send+sync"
+
+[[object]]
+name = "GLib.Win32OSType"
+status = "generate"
+cfg_condition = "windows"
