+++ glib/src/auto/enums.rs
@@ -2352,3 +2352,49 @@ impl FromGlib<ffi::GVariantClass> for VariantClass {
         }
     }
 }
+
+#[cfg(windows)]
+#[cfg_attr(docsrs, doc(cfg(windows)))]
+#[derive(Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Clone, Copy)]
+#[non_exhaustive]
+#[doc(alias = "GWin32OSType")]
+pub enum Win32OSType {
+    #[doc(alias = "G_WIN32_OS_ANY")]
+    Any,
+    #[doc(alias = "G_WIN32_OS_WORKSTATION")]
+    Workstation,
+    #[doc(alias = "G_WIN32_OS_SERVER")]
+    Server,
+    #[doc(hidden)]
+    __Unknown(i32),
+}
+
+#[cfg(windows)]
+#[doc(hidden)]
+impl IntoGlib for Win32OSType {
+    type GlibType = ffi::GWin32OSType;
+
+    #[inline]
+    fn into_glib(self) -> ffi::GWin32OSType {
+        match self {
+            Self::Any => ffi::G_WIN32_OS_ANY,
+            Self::Workstation => ffi::G_WIN32_OS_WORKSTATION,
+            Self::Server => ffi::G_WIN32_OS_SERVER,
+            Self::__Unknown(value) => value,
+        }
+    }
+}
+
+#[cfg(windows)]
+#[doc(hidden)]
+impl FromGlib<ffi::GWin32OSType> for Win32OSType {
+    #[inline]
+    unsafe fn from_glib(value: ffi::GWin32OSType) -> Self {
+        match value {
+            ffi::G_WIN32_OS_ANY => Self::Any,
+            ffi::G_WIN32_OS_WORKSTATION => Self::Workstation,
+            ffi::G_WIN32_OS_SERVER => Self::Server,
+            value => Self::__Unknown(value),
+        }
+    }
+}
