+++ glib/sys/Gir.toml
@@ -69,34 +69,34 @@ name = "GLib.*"
 status = "generate"
     [[object.function]]
     name = "win32_locale_filename_from_utf8"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_get_windows_version"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_getlocale"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_check_windows_version"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_error_message"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_ftruncate"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_get_command_line"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_get_package_installation_directory"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_get_package_installation_directory_of_module"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "win32_get_package_installation_subdirectory"
-    cfg_condition = "window"
+    cfg_condition = "windows"
     [[object.function]]
     name = "unix_get_passwd_entry"
     cfg_condition = "unix"
@@ -138,7 +138,7 @@ status = "ignore"
 [[object]]
 name = "GLib.Win32OSType"
 status = "generate"
-cfg_condition = "window"
+cfg_condition = "windows"
 
 [[object]]
 name = "GLib.macro__has_attribute___noreturn__"
